dashboard/requirements.txt
altair==5.3.0
asn1crypto==1.5.1
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.34.108
botocore==1.34.108
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
protobuf==4.25.3
pyarrow==16.1.0
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redshift-connector==2.1.1
referencing==0.35.1
requests==2.31.0
rich==13.7.1
rpds-py==0.18.1
s3transfer==0.10.1
scramp==1.4.5
setuptools==69.5.1
six==1.16.0
smmap==5.0.1
soupsieve==2.5
streamlit==1.34.0
tenacity==8.3.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
watchdog==4.0.0
